+++ src/launchpad_s/mod.rs
@@ -12,7 +12,7 @@ pub struct Spec;
 impl crate::DeviceSpec for Spec {
     const BOUNDING_BOX_WIDTH: u32 = 9;
 	const BOUNDING_BOX_HEIGHT: u32 = 9;
-	const COLOR_PRECISION: u8 = 4;
+	const COLOR_PRECISION: u16 = 4;
 	
     type Input = LaunchpadSInput;
 	type Output = LaunchpadSOutput;
